Work on #1852 

Implementing a suggestion from @essweine - pass in the last completed spiff task when getting the next round of ready engine steps. If none are returned try once from the task root to make sure the next task is not on a completely different branch. This simulates the deeper library improvements that are the queue, but the impact is noticeable. Measured with 150 iterations, run time is cut by ~45% and the number of function calls are cut by ~50%. Calls to `_next` are also proportional to increases in iteration sizes (before a 3x in iteration size caused ~8x increase in calls to `_next`).

I also modified the way the profiling is enabled, instead of having to add an import and switch the `WorkflowExecutionService` to `ProfiledWorkflowExecutionService`, you can just pass in `profile=True` to `run_and_save`.

Before:

```
         12238463 function calls (11908765 primitive calls) in 32.884 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     1830    0.049    0.000   29.148    0.016 workflow.py:99(get_tasks)
     9401    1.222    0.000   24.551    0.003 task.py:220(__next__)
   591001    5.170    0.000   22.982    0.000 task.py:55(_next)

```

After:

```
         6806697 function calls (6478600 primitive calls) in 18.292 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     1217    0.023    0.000    8.584    0.007 workflow.py:99(get_tasks)
     8175    0.639    0.000    9.647    0.001 task.py:220(__next__)
   259650    2.301    0.000    9.316    0.000 task.py:55(_next)

```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
  - Introduced performance profiling option in the `run_and_save` method to help monitor and improve workflow execution performance.
  
- **Refactor**
  - Enhanced task filtering logic to use `TaskFilter` for more efficient and streamlined task management.
  
- **Bug Fixes**
  - Improved internal method naming conventions for better code clarity and maintainability.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->